+++ back/src/main/java/org/pknu/weather/domain/Member.java
@@ -1,8 +1,8 @@
-package org.pknu.weather.member.domain;
+package org.pknu.weather.domain;
 import jakarta.persistence.*;
 import lombok.*;
 import org.pknu.weather.common.BaseEntity;
-import org.pknu.weather.location.domain.Location;
+import org.pknu.weather.domain.common.Sensitivity;
 
 @Entity
 @Getter
