+++ back/src/main/java/org/pknu/weather/domain/Post.java
@@ -1,10 +1,7 @@
-package org.pknu.weather.post.domain;
+package org.pknu.weather.domain;
 import jakarta.persistence.*;
 import lombok.*;
 import org.pknu.weather.common.BaseEntity;
-import org.pknu.weather.location.domain.Location;
-import org.pknu.weather.member.domain.Member;
-import org.pknu.weather.tag.domain.Tag;
 
 @Entity
 @Getter
